+++ CHANGELOG.md
@@ -1,4 +1,20 @@
 # 0.4.4 (unpublished)
+- Big reorganization of `fidget::raster`
+    - `render2d` and `render3d` modules are renamed to `pixel` and `voxel`,
+      respectively
+    - `ImageRenderConfig` and `VoxelRenderConfig` are both renamed to
+      `RenderConfig` (in different namespaces)
+    - Similarly, each namespace contains an `Image` and `RenderSize` type
+      definition.  `voxel::Image` replaces `GeometryBuffer`; `pixel::Image` is
+      equivalent to `fidget::raster::Image<DistancePixel>` in the previous code
+      (but see the next point about renaming!)
+    - Renamed `DistancePixel` to `RawDistancePixel`; added an unpacked `enum
+      DistancePixel` (instead of using `Result<f32, PixelFill>` as a weird
+      `Either` type
+    - The tile sizes member of `RenderConfig` is now optional and will fall back
+      to the evaluator's tile size if not populated
+        - This adds a `F: RenderHints` trait bound for many functions in
+          `fidget::raster`
 - Add `Image::build` function to build an image from a `Vec<T>` and
   `ImageSizeLike`, returning an error if the data size is incorrect.
 - Revamping `fidget-bytecode`